[REFACTOR] 코드 타입에 맞춰 CrawlingServiceImpl 추가
